feat(hexagon-rt): implement command buffer finalization (#974)
The finalization of a command buffer means that all dispatches and execution barriers have been added to it and it can be prepared to be executed on the device. In the case of Hexagon command buffers, the dispatches and execution barrier entries are collected on the ARM side before finalization. Thus, during finalization, the information about all entries needs to be transferred to the device. As the DSP has different type sizes and alignments, the command buffer is serialized and the serialized buffer is handed over to the DSP side, where it is stored.

As the serialization is non-trivial and can be decoupled from the runtime driver and its dependencies of actually running on ARM/DSP at the same time, it can be unit tested. In order to build and run the parts of the Hexagon driver (only the serialization for now) independently from the runtime, this PR introduces a new sub-library named "units", which contains all the parts that can be unit tested without the runtime.