feat(hexagon-rt): add support for buffer copies/transfers to command buffer, implement serialize bindings (#1010)
Implement support for adding buffer copy/transfer operations to the Hexagon command buffers.
Extend the serialization of command buffers to support this as well.

Implement serializing the buffer bindings used to supply actual buffers to command buffer executions.
This is a main step towards ROO-878 (Execute Command Buffers).

Minor extensions and fixes in this PR are:
- Make command buffers retain resources (buffers) used inside, so those don't get destroyed prematurely.
- RPC memory is cached and coherent on the host by default, so enable RPC memory buffer allocation for cached/coherent host buffers that are supposed to be device visible (it failed before, reporting no memory type supporting this)
- Add DSP addresses to serialized command buffers.
- Update the buffer serialization unit tests: add test of binding serialization, add test of copy/transfer operation, check DSP addresses