fix(hexagon): add query function during serialization (#1014)
### TL;DR

Add query function to Hexagon executable for function lookup at runtime. Additionally, did some refactoring to cleanup the code.

### What changed?

- Added query function generation for Hexagon target to enable the HAL to call functions by their ordinals
- Added a new test to verify query function generation and visibility in the shared object
- Refactored serialization into smaller, more focused methods
- Changed test examples from f32 to i32 tensor types for easier debugging when executing

### How to test?

Run the new test that verifies the query function is properly generated:

```
bazel test @patio//...
```

Full testing for correctness must be done with Stefan's open prs for the runtime that load the generated executable. We have tested this together and will be adding this to the CI in a close future...