feat(hexagon-rt): support workgroups (#1054)
Kernel dispatches can use multiple workgroups, which means the same kernel is
to be executed with the same buffers multiple times with different workgroups
ids. Basically, it means the kernel is executed in a 3-dimensional loop nest
with the loop indices being passed into the kernel as parameters.

This PR adds support for tracking the 3 dimensional work group parameters from
the command buffer recording to the actual command buffer execution on the DSP
and add the loop nest there.