test: correct non deterministic test (#1065)
Fix non determinism that appeared after the iree bump and went unnoticed. The printed pass is called for each dispatch and depending on the compilation order of the dispatches, this test could previously fail. The important part of the test is the linalg.mmt4d, so it is ok to remove the pack and unpack.