feat(hexagon-rt): support constants (#1055)
Kernel dispatches can pass constants to the kernel. This PR adds support for
this. This means storing the constants passed to adding the dispatch during
command buffer recording in the command buffer, serializing those for passing
those to the DSP and finally passing the pointer to the constants to the
kernel running on the DSP.