refactor(hexagon-rt): avoid duplication of command buffer data structures (#1077)
Refactor the command buffer to store serialized commands in the linked list
instead of storing the commands in dedicated data structures and serializing
everything on finalization. This makes the serialization during finalization
to become a concatenation of the already serialized commands. Thus, it becomes
much simpler. This change also avoids having native storage data structures
in parallel to the serialized data structures.